requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.32.0
jinja2==3.1.2
python-multipart==0.0.6
pydantic>=2.7.4

# LangChain & LangGraph
langchain==0.3.12
langchain-openai==0.2.14
langchain-core==0.3.28
langgraph==0.2.60
langsmith==0.2.6

# Search Tools
tavily-python==0.5.0
duckduckgo-search==6.3.5

# Utilities
python-dotenv==1.0.0
aiofiles==24.1.0

# Caching
redis==5.0.1
hiredis==2.3.2

